Adding a test when cloning to ensure something is checked out
Previously, after a repo was cloned, it was assumed that the branch
specified in the config file was in fact a branch. If this happened to
be a hash instead, a checkout of origin/hash was attempted which would
fail.

This commit allows a hash to be specified, and successfully checks it
out if possible.