Redo run-with-restart function
It appears the run with restart function was never being called because
there was a stray space and lettuce was matching steps that called for
it to the standard run function.  After fixing the stray space, I also
found a number of issues with how the restart function was operating.

I have largely rewritten it.  It now assumes that the previous run was
set up to:
  * be 2 time steps long
  * have output interval set to the timestep length
  * have restart interval set to the timestep length

Then the function does a cold start that runs for one time step, then a
restart that runs for one time step more.

This could of course be generalized, but this seems like minimally
useful way to do a restart test.