Fix namelist setting, and add directory when errors occur
This commit fixes the step that can be used to set a namelist option.
Previously, if there were other namelist options that partially matched
the option that was being set, they would be set as well.

Additionally, this commit adds a print of the directory name when an
error occurs to help with debugging errors.