Turn subprocess check_call to call when running MPAS
The current version of Albany on blueskies dies on finalization, making
lettuce think the model failed.  But those runs are actually ok.  This
commit changes subprocess.check_call to call so that the return code of
MPAS will not be checked.  This is somewhat hazardous as we won't be
ablt to detect actual aborts.  But hopefully those situations will lead
to the test failing for other reasons.